+++ src/aux-common/bots/BotEvents.ts
@@ -204,7 +204,8 @@ export type AsyncActions =
     | WatchLoomAction
     | GetLoomMetadataAction
     | GetScriptIssuesAction
-    | LoadSharedDocumentAction;
+    | LoadSharedDocumentAction
+    | CalculateScreenCoordinatesFromPositionAction;
 
 export type RemoteBotActions =
     | GetRemoteCountAction
@@ -3488,6 +3489,27 @@ export interface CalculateScreenCoordinatesFromViewportCoordinatesAction
 }
 
 /**
+ * Defines an event that calculates the 2D screen coordinates from the given 3D positions.
+ *
+ * @dochash types/os/portals
+ * @docname CalculateScreenCoordinatesFromPositionAction
+ */
+export interface CalculateScreenCoordinatesFromPositionAction
+    extends AsyncAction {
+    type: 'calculate_screen_coordinates_from_position';
+
+    /**
+     * The portal that the ray should be calculated for.
+     */
+    portal: CameraPortal;
+
+    /**
+     * The 3D positions that the screen coordinates should be calculated for.
+     */
+    coordinates: Point3D[];
+}
+
+/**
  * Defines an event that calculates the 2D viewport coordinates from the given 2D screen coordinates.
  *
  * @dochash types/os/portals
@@ -5911,6 +5933,19 @@ export function calculateViewportCoordinatesFromScreenCoordinates(
 ): CalculateViewportCoordinatesFromScreenCoordinatesAction {
     return {
         type: 'calculate_viewport_coordinates_from_screen_coordinates',
+        portal,
+        coordinates,
+        taskId,
+    };
+}
+
+export function calculateScreenCoordinatesFromPosition(
+    portal: CameraPortal,
+    coordinates: Point3D[],
+    taskId?: number | string,
+): CalculateScreenCoordinatesFromPositionAction {
+    return {
+        type: 'calculate_screen_coordinates_from_position',
         portal,
         coordinates,
         taskId,