+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 -   Added the `Sec-Websocket-Protocol=casualos.records` header for websocket requests made to the records system.
     -   This can help load balancers differentiate between records requests and inst requests.
 -   Added a better date of birth input to the "Sign Up" page.
+-   Added the `os.calculateScreenCoordinatesFromPosition()` function.
+-   Added batching for events sent from the main thread to the worker thread.
 
 ### :bug: Bug Fixes